requirements.txt
paho-mqtt
pandas
scikit-learn
joblib
numpy
